Mv _Scales to dimensions.Scales
